+++ src/Sql/Sql/Data Sync/Model/AzureSqlSyncGroupModel.cs
@@ -14,7 +14,7 @@
 
 using System;
 using System.Security;
-using Microsoft.Azure.Management.Sql.LegacySdk.Models;
+using Microsoft.Azure.Management.Sql.Models;
 
 namespace Microsoft.Azure.Commands.Sql.DataSync.Model
 {
@@ -89,6 +89,11 @@ namespace Microsoft.Azure.Commands.Sql.DataSync.Model
         public AzureSqlSyncGroupSchemaModel Schema { get; set; }
 
         /// <summary>
+        /// Gets or sets if private link connection should be used
+        /// </summary>
+        public bool? UsePrivateLinkConnection { get; set; }
+
+        /// <summary>
         /// Construct AzureSqlSyncGroupModel
         /// </summary>
         public AzureSqlSyncGroupModel()
@@ -110,13 +115,14 @@ namespace Microsoft.Azure.Commands.Sql.DataSync.Model
             DatabaseName = databaseName;
             ResourceId = syncGroup.Id;
             SyncGroupName = syncGroup.Name;
-            IntervalInSeconds = syncGroup.Properties.Interval;
-            SyncDatabaseId = syncGroup.Properties.SyncDatabaseId;
-            HubDatabaseUserName = syncGroup.Properties.HubDatabaseUserName;
-            ConflictResolutionPolicy = syncGroup.Properties.ConflictResolutionPolicy == null ? null : syncGroup.Properties.ConflictResolutionPolicy.ToString();
-            SyncState = syncGroup.Properties.SyncState;
-            LastSyncTime = syncGroup.Properties.LastSyncTime;
-            Schema = syncGroup.Properties.Schema == null ? null : new AzureSqlSyncGroupSchemaModel(syncGroup.Properties.Schema);
+            IntervalInSeconds = syncGroup.Interval;
+            SyncDatabaseId = syncGroup.SyncDatabaseId;
+            HubDatabaseUserName = syncGroup.HubDatabaseUserName;
+            ConflictResolutionPolicy = syncGroup.ConflictResolutionPolicy == null ? null : syncGroup.ConflictResolutionPolicy.ToString();
+            SyncState = syncGroup.SyncState;
+            LastSyncTime = syncGroup.LastSyncTime;
+            Schema = syncGroup.Schema == null ? null : new AzureSqlSyncGroupSchemaModel(syncGroup.Schema);
+            UsePrivateLinkConnection = syncGroup.UsePrivateLinkConnection;
         }
     }
 }