+++ src/DeviceProvisioningServices/DeviceProvisioningServices/ChangeLog.md
@@ -18,6 +18,15 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
-
+* Manage Device Enrollments. New cmdlets are:
+    - `Add-AzIoTDeviceProvisioningServiceEnrollment`
+    - `Get-AzIoTDeviceProvisioningServiceEnrollment`
+    - `Remove-AzIoTDeviceProvisioningServiceEnrollment`
+    - `Set-AzIoTDeviceProvisioningServiceEnrollment`
+* Manage Enrollment Groups. New cmdlets are:
+    - `Add-AzIoTDeviceProvisioningServiceEnrollmentGroup`
+    - `Get-AzIoTDeviceProvisioningServiceEnrollmentGroup`
+    - `Remove-AzIoTDeviceProvisioningServiceEnrollmentGroup`
+    - `Set-AzIoTDeviceProvisioningServiceEnrollmentGroup`
 ## Version 0.7.2
 * Update references in .psd1 to use relative path